refactor: option-button 컴포넌트화
